Retira regra de nogocio do input para robo proprio e tira metodos da main
