+++ robots/input.js
@@ -1,0 +1,27 @@
+const readline = require('readline-sync')
+const state = require('./state.js')
+
+function robot() {
+    const content = {
+        maximumSentences: 7
+    }
+
+    content.searchTerm = askAndReturnSearchTerm()
+    content.prefix = askAndReturnPrefix()
+    state.save(content)
+
+    function askAndReturnSearchTerm() {
+        return readline.question('Type a Wikipedia search term: ')
+    }
+
+    function askAndReturnPrefix() {
+        const prefixes = ['Who is', 'What is', 'The history of']
+        //Pega o index do array selecionado pelo usuario
+        const seletedPrefixIndex = readline.keyInSelect(prefixes, 'Choose one option: ')
+        const selectedPrefixText = prefixes[seletedPrefixIndex]
+
+        return selectedPrefixText;
+    }
+}
+
+module.exports = robot
